ilasm: Some work on supporting native types and marshaling info on fields.
A bunch of native/variant types aren't supported. These are prett much
all deprecated. Microsoft's ILAsm seems to still emit them correctly,
albeit with a warning, though. In this regard, we're incompatible, but
I think this is a non-issue as no one should be using these types in
this day and age. (See the variant_type and native_type grammar rules.)